# api/openapi/atlas-v1.yaml
openapi: 3.1.0
info:
  title: Atlas API
  version: 0.0.8
  description: >
    Atlas REST API contract. Covers local health, the cluster index and
    cluster-scoped inventory reads (health, OSDs, Hosts, Storage
    Devices, Ceph Daemons, Pools), cluster registration and Agent
    enrollment, Agent observation pushes, explicit inventory sync run
    history, case records with cluster filtering, and the authenticated
    operator identity probe. Write endpoints and cluster-registration
    management carry a bearer-token security scheme; inventory read
    endpoints remain public. 0.0.8 is a
    documented 0.x breaking change: it adds the cluster index and
    per-cluster reads, and removes the single-cluster
    /api/v1/clusters/current/* family.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
paths:
...
  /api/v1/me:
    get:
      tags:
        - Identity
      operationId: getMe
      summary: Get the authenticated operator identity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Identity of the operator whose bearer token was verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
              examples:
                operator:
                  value:
                    subject: operator-1
                    displayName: Storage Operator
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    error:
                      class: Unauthorized
                      message: missing bearer token
  /api/v1/clusters:
    get:
      tags:
        - Clusters
      operationId: listClusters
      summary: List clusters
      description: >
        The cluster index: every active registered cluster with its
        latest observed health and Agent last-seen time. Searchable by
        name or FSID, paginated.
      parameters:
        - name: q
          in: query
          description: Case-insensitive substring match against cluster name or FSID.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size, clamped to at most 100; 0 or absent uses the default 50.
          schema:
            type: integer
            minimum: 0
        - name: offset
          in: query
          description: Page offset; defaults to 0.
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: One page of the cluster index plus the total the query matches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterIndex"
        "400":
          description: limit or offset is not a non-negative integer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Clusters
      operationId: createClusterRegistration
      summary: Register a cluster
      description: >
        Registers a Cluster (ADR-0025) and returns the one-time Enrollment
        Credential exactly once. The credential is stored only as a hash,
        expires if unused, and is never returned by any other endpoint.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClusterRegistrationRequest"
            examples:
              prodCluster:
                value:
                  name: prod-ceph-a
                  clusterType: bare-metal
      responses:
        "201":
          description: Registered cluster with its one-time enrollment credential.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClusterRegistrationResponse"
        "400":
          description: Request body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    error:
                      class: Unauthorized
                      message: missing bearer token
        "422":
          description: Cluster registration requires the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/clusters/{id}:
    get:
      tags:
        - Clusters
      operationId: getClusterRegistration
      summary: Get a cluster registration
      description: >
        Addresses the cluster by registration id rather than FSID: the FSID
        is nil from registration until Enrollment binds it (ADR-0026), so
        FSID-addressed management of fresh registrations is impossible.
        Registration management is authenticated like the writes it
        reports on; inventory reads stay public until RBAC (v0.0.9).
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Cluster registration id.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Cluster registration detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterRegistration"
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  value:
                    error:
                      class: Unauthorized
                      message: missing bearer token
        "404":
          description: No cluster registration exists with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Cluster registration requires the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Clusters
      operationId: deregisterCluster
      summary: Deregister a cluster
      description: >
        Retires the registration and consumes any live Enrollment
        Credential. Snapshots and Cases are preserved untouched.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Cluster registration id.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Deregistered cluster registration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterRegistration"
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No cluster registration exists with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The cluster is already deregistered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Cluster registration requires the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/agent/enroll:
    post:
      tags:
        - Agents
      operationId: enrollAgent
      summary: Enroll an Atlas Agent
      description: >
        Runs the enrollment handshake (ADR-0026): an Atlas Agent presents a
        certificate signing request, the one-time Enrollment Credential, and
        the FSID it self-reported from its first Dashboard read. Atlas burns
        the credential, binds the FSID to the registered Cluster, and issues
        a client certificate chained to the Atlas internal CA. The
        credential in the body is the authentication — this is the one write
        endpoint that does not take a bearer token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentEnrollmentRequest"
      responses:
        "201":
          description: Enrolled cluster with the issued certificate chain.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentEnrollmentResponse"
        "400":
          description: Request body failed validation or the CSR is malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: The enrollment credential is invalid, expired, or already used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The FSID is already bound to another registered cluster.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Agent enrollment requires the enrollment CA and the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/agent/observations:
    post:
      tags:
        - Agents
      operationId: pushAgentObservations
      summary: Push an Atlas Agent observation batch
      description: >
        Accepts one typed Observation Batch for one collection cycle
        (ADR-0025): all entity types, one consistent snapshot, the same
        normalized shape the sync command produces. Authentication is the
        enrolled client certificate over mutual TLS; Atlas attributes the
        batch to the Cluster the certificate maps to by its recorded
        serial number and rejects a batch whose FSID does not match
        (ADR-0026). The batch is atomic — it fully persists or fails, and
        Atlas records an inventory sync run with provider "agent". The
        request deliberately carries no provider or scenario: Atlas
        records the provider server-side.
      security:
        - agentCertificate: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentObservationBatch"
      responses:
        "201":
          description: The batch persisted as one snapshot; cluster attribution came from the certificate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentObservationReceipt"
        "400":
          description: Request body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No verified client certificate was presented, or the certificate is unknown, revoked, expired, or attached to a deregistered cluster.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The batch's cluster FSID does not match the cluster the certificate is enrolled for.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Agent observations require the enrollment CA and the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/inventory-sync-runs:
    get:
      tags:
        - Inventory
      operationId: listInventorySyncRuns
      summary: List recent inventory sync runs
      parameters:
        - name: cluster
          in: query
          description: Optional cluster FSID; only that cluster's runs list.
            Runs the history cannot attribute to a cluster — failed pulls
            that never reached one — never appear in a scoped list.
          schema:
            type: string
      responses:
        "200":
          description: Recent inventory sync run history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventorySyncRun"
              examples:
                runs:
                  value:
                    - id: 2
                      provider: fake
                      scenario: missing
                      status: failed
                      startedAt: "2026-08-13T12:01:00Z"
                      finishedAt: "2026-08-13T12:01:01Z"
                      errorClass: Unavailable
                      errorMessage: "Unavailable: read fixture"
                    - id: 1
                      provider: fake
                      scenario: reef-osd-down-baremetal
                      status: succeeded
                      startedAt: "2026-08-13T12:00:00Z"
                      finishedAt: "2026-08-13T12:00:01Z"
                      snapshotId: 1
                      clusterFsid: 00000000-0000-4000-8000-000000000102
                      clusterName: reef-fixture
        default:
          $ref: "#/components/responses/Error"
  /api/v1/alert-evaluation-runs:
    get:
      tags:
        - Inventory
      operationId: listAlertEvaluationRuns
      summary: List recent alert evaluation runs
      responses:
        "200":
          description: Recent alert evaluation run history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertEvaluationRun"
              examples:
                runs:
                  value:
                    - id: 2
                      provider: fake
                      scenario: osd-down-alert
                      status: succeeded
                      startedAt: "2026-08-14T09:20:00Z"
                      finishedAt: "2026-08-14T09:20:01Z"
                      alertsEvaluated: 1
                      casesCreated: 1
                    - id: 1
                      provider: fake
                      scenario: provider-unauthorized
                      status: failed
                      startedAt: "2026-08-14T09:15:00Z"
                      finishedAt: "2026-08-14T09:15:01Z"
                      errorClass: Unauthorized
                      errorMessage: "Unauthorized: simulated observability source rejected provider credentials"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/cases:
    get:
      tags:
        - Cases
      operationId: listCases
      summary: List cases
      parameters:
        - name: cluster
          in: query
          description: Optional cluster FSID; only that cluster's cases list.
          schema:
            type: string
      responses:
        "200":
          description: Recent cases ordered by most recent update.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
              examples:
                seededCases:
                  value:
                    - id: 2
                      title: Review weekly capacity trend
                      summary: Local seed case for capacity review without an automation workflow attached.
                      status: triaged
                      severity: medium
                      source: manual
                      clusterFsid: 00000000-0000-4000-8000-000000000101
                      createdAt: "2026-08-13T12:05:00Z"
                      updatedAt: "2026-08-13T12:10:00Z"
                    - id: 1
                      title: OSD down requires triage
                      summary: Local seed case representing an OSD_DOWN health check that needs operator triage.
                      status: detected
                      severity: high
                      source: prometheus
                      clusterFsid: 00000000-0000-4000-8000-000000000101
                      createdAt: "2026-08-13T12:00:00Z"
                      updatedAt: "2026-08-13T12:00:00Z"
        "400":
          description: The cluster filter is not a UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Cases
      operationId: createCase
      summary: Create a manual case
      description: >
        Creates a Case in the detected status with source forced to manual and
        records a case_detected Timeline Event attributed to the verified
        operator.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
            examples:
              capacityReview:
                value:
                  title: Manual review of slow OSD warnings
                  summary: Operator noticed slow OSD alerts outside automated detection.
                  severity: medium
                  clusterFsid: 00000000-0000-4000-8000-000000000101
      responses:
        "201":
          description: Created case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "400":
          description: Request body failed validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Case writes require the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/cases/{id}:
    get:
      tags:
        - Cases
      operationId: getCase
      summary: Get a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Case detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/cases/{id}/timeline:
    get:
      tags:
        - Cases
      operationId: listCaseTimeline
      summary: List a case timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Case Timeline Events ordered oldest to newest.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
              examples:
                seededTimeline:
                  value:
                    - id: 2
                      caseId: 2
                      type: case_detected
                      message: Manual capacity review case created.
                      occurredAt: "2026-08-13T12:05:00Z"
                      actor:
                        type: user
                        displayName: Storage Operator
                      payload:
                        source: manual
                        clusterFsid: 00000000-0000-4000-8000-000000000101
                    - id: 3
                      caseId: 2
                      type: case_triaged
                      message: Capacity review case triaged.
                      occurredAt: "2026-08-13T12:10:00Z"
                      actor:
                        type: user
                        displayName: Storage Operator
                      payload:
                        previousStatus: detected
                        newStatus: triaged
        default:
          $ref: "#/components/responses/Error"
  /api/v1/cases/{id}/transitions:
    post:
      tags:
        - Cases
      operationId: transitionCase
      summary: Transition a case status
      description: >
        Applies a manual lifecycle transition and records the matching
        Timeline Event (case_triaged or case_status_changed). Allowed edges
        are detected -> triaged, detected -> closed, and triaged -> closed;
        closed is terminal and reopening means creating a new case.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionCaseRequest"
            examples:
              triage:
                value:
                  status: triaged
      responses:
        "200":
          description: Updated case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "400":
          description: Unknown target status or malformed body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No bearer token was presented, the token failed verification, or authentication is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Case not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The requested transition is not allowed from the current status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Case writes require the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/workflow-instances/{id}/jobs:
    get:
      tags:
        - Cases
      operationId: listWorkflowJobs
      summary: List a Workflow Instance's Jobs
      description: >
        The instance's Jobs in definition order, with per-Job state and
        attempt progress (ADR-0019). Job transitions do not write Timeline
        Events; instance-level state changes do.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Jobs ordered by definition position.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowJob"
        "404":
          description: Workflow Instance not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Workflow reads require the postgres read source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          $ref: "#/components/responses/Error"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        RS256 JWT issued by the configured OIDC issuer, verified against its
        JWKS. Required on Operator write endpoints; read endpoints stay
        public. The agent endpoints are the exceptions — enrollment
        authenticates with its one-time enrollment credential in the body,
        and observation pushes authenticate with the enrolled client
        certificate over mutual TLS (ADR-0026).
    agentCertificate:
      type: mutualTLS
      description: >
        Client certificate chained to the Atlas internal enrollment CA
        (ADR-0026), presented over mutual TLS on the API's TLS listener.
        Atlas maps the certificate serial to exactly one registered
        Cluster; the certificate is the authentication and the cluster
        attribution, and nothing is accepted from payload claims.
  responses:
    Error:
      description: Request failed before a valid resource response could be produced.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            providerUnavailable:
              value:
                error:
                  class: Unavailable
                  message: provider is unavailable
  schemas:
...
    MeResponse:
      type: object
      additionalProperties: false
      required:
        - subject
        - displayName
      properties:
        subject:
          type: string
          description: Verified OIDC subject of the operator.
        displayName:
          type: string
          description: Display name resolved from token claims, falling back to the subject.
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...